+++ include/cask/deferred/PromiseDeferred.hpp
@@ -29,6 +29,7 @@ public:
     void onSuccess(std::function<void(T)> callback) override;
     void onError(std::function<void(E)> callback) override;
     void onCancel(const std::function<void()>& callback) override;
+    void onShutdown(const std::function<void()>& callback) override;
     void cancel() override;
     T await() override;
 
@@ -69,6 +70,11 @@ void PromiseDeferred<T,E>::onError(std::function<void(E)> callback) {
 template <class T, class E>
 void PromiseDeferred<T,E>::onCancel(const std::function<void()>& callback) {
     promise->onCancel(callback);
+}
+
+template <class T, class E>
+void PromiseDeferred<T,E>::onShutdown(const std::function<void()>& callback) {
+    promise->onShutdown(callback);
 }
 
 template <class T, class E>