+++ test/cask/scheduler/TestSingleThreadScheduler.cpp
@@ -139,3 +139,57 @@ TEST_F(SingleThreadSchedulerTest, RegistersCallbackAfterCancelled) {
     EXPECT_EQ(cancel_counter, 1);
     awaitIdle();
 }
+
+TEST_F(SingleThreadSchedulerTest, RunsShutdownCallbackAfterTimerTaskCompletion) {
+    bool shutdown = false;
+    std::mutex mutex;
+    mutex.lock();
+
+    auto before = std::chrono::high_resolution_clock::now();
+    auto cancelable = sched->submitAfter(25, [&mutex] {
+        mutex.unlock();
+    });
+
+    cancelable->onShutdown([&shutdown] {
+        shutdown = true;
+    });
+
+    mutex.lock();
+    auto after = std::chrono::high_resolution_clock::now();
+
+    auto delta = after - before;
+    auto milliseconds = std::chrono::duration_cast<std::chrono::milliseconds>(delta).count();
+
+    EXPECT_GE(milliseconds, 25);
+    EXPECT_TRUE(shutdown);
+    
+    awaitIdle();
+}
+
+TEST_F(SingleThreadSchedulerTest, RunsShutdownImmediatelyCallbackIfTimerAlreadyFired) {
+    bool shutdown = false;
+    std::mutex mutex;
+    mutex.lock();
+
+    auto before = std::chrono::high_resolution_clock::now();
+    auto cancelable = sched->submitAfter(25, [&mutex] {
+        mutex.unlock();
+    });
+
+    mutex.lock();
+    auto after = std::chrono::high_resolution_clock::now();
+
+    auto delta = after - before;
+    auto milliseconds = std::chrono::duration_cast<std::chrono::milliseconds>(delta).count();
+
+    EXPECT_GE(milliseconds, 25);
+    EXPECT_FALSE(shutdown);
+
+    cancelable->onShutdown([&shutdown] {
+        shutdown = true;
+    });
+
+    EXPECT_TRUE(shutdown);
+    
+    awaitIdle();
+}